Commented code

+++ src/Generator.java
@@ -18,28 +18,29 @@ public class Generator {
 		}
 		
 		ArrayList<Color> colors = new ArrayList<Color>(MAX_COLOR);
-		/* Palette is a color scheme. It iterates from the first rgb color
+		/* 'palette' is a color scheme. It iterates from the first rgb color
 		 * and adds the colors in between to the ArrayList called 'colors'
 		 */
-		Color[] palette = { new Color(255, 0, 0), // The color scheme
-							new Color(0, 255, 0 ),
-							new Color(0, 0, 255 ),
-							new Color(255, 255, 0),
-							new Color(0,0,0)};
+		Color[] palette = { new Color(255, 0, 0), 	// RED
+							new Color(0, 255, 0 ), 	// GREEN
+							new Color(0, 0, 255 ), 	// BLUE
+							new Color(255, 255, 0), // YELLOW
+							new Color(0,0,0)};		// BLACK
 	
 		/* The color will iterate 'c_iterations' amount of times before reaching the next color.
-		 * Example: 255(3-1) ~= 127. It will go change its value 127 times from 255, 0, 0 before reaching the next color 0, 255, 0
+		 * Example: 255(5-1) = 63.75. It will go change its value 127 times from 255, 0, 0 before reaching the next color 0, 255, 0
 		*/
-		double c_iterations = MAX_COLOR/(palette.length-1); 
+		double c_iterations = (double)MAX_COLOR/(palette.length-1); 
 		for (int j = 0; j < palette.length - 1; j++) {
 			Color c0 = palette[j],
 				  c1 = palette[j+1];
-			// The red, green and blue changefactors from one color to another.
+			// The red, green and blue change-factors from one color to the next.
 			double dr = (c1.getRed() - c0.getRed())/c_iterations,
 				   dg = (c1.getGreen() - c0.getGreen())/c_iterations,
 				   db = (c1.getBlue() - c0.getBlue())/c_iterations;
+			
+			// Add 'c_iteration' amount of colors to the 'colors'-List. Start with the old color and gradually change it to the new color.
 			for (int i = 0; i < c_iterations; i++) {
-				// Add 'c_iteration' amount of colors to the 'colors' List. Start with the old color and gradually change it to the new color.
 				colors.add(new Color(c0.getRed() + (int)(i*dr), c0.getGreen() + (int)(i*dg), c0.getBlue() + (int)(i*db)));
 			}
 		}
@@ -77,11 +78,16 @@ public class Generator {
 		gui.enableInput();
 	}
 	
+	// Generate and returns a plane of complex numbers.
 	private Complex[][] mesh(double minRe, double maxRe, double minIm, double maxIm, int width, int height) {
 		Complex[][] complex = new Complex[height][width];
+		// Calculate the change-factor per pixel for the real & imaginary part of the complex plane.
 		double rePart = (maxRe - minRe) / width;
 		double imPart = (maxIm - minIm) / height;
 		
+		/* Start with the upper-left complex number in the plane and gradually change it to the lower-right complex number.
+		 * Add all complex numbers for each pixel in between.
+		*/
 		for (int y = 0; y < height; y++) {
 			for (int x = 0; x < width; x++) {
 				complex[y][x] = new Complex(minRe + rePart * x, maxIm - imPart * y);
